Rename file for logging.
