+++ src/rt/verona.h
@@ -23,6 +23,6 @@
 #include "sched/multimessage.h"
 #include "sched/noticeboard.h"
 #include "sched/schedulerthread.h"
-#include "test/systematic.h"
+#include "test/logging.h"
 
 #include <snmalloc.h>
